Bug fix and needed feature implement
Fixed a crash that would occur when clearing fields for a file that had
not been saved for the day. Fixed the input field check for different
scenarios. Added a save ability when exiting the program (in case it
was forgotten.) Preview window will now disable buttons in main screen,
so no "hidden window" issue. Updated README for issue resolution.